--- app/src/main/java/com/example/taskyapplication/agenda/task/data/network/models/TaskRemoteDataSource.kt
@@ -1,29 +1,0 @@
-package com.example.taskyapplication.agenda.task.data.network.models
-
-import com.example.taskyapplication.agenda.task.domain.RemoteDataSource
-import com.example.taskyapplication.agenda.task.domain.model.GetTaskResponse
-import com.example.taskyapplication.agenda.task.domain.network.TaskApiService
-import com.example.taskyapplication.domain.utils.DataError
-import com.example.taskyapplication.domain.utils.EmptyResult
-import com.example.taskyapplication.domain.utils.Result
-import javax.inject.Inject
-
-class TaskRemoteDataSource @Inject constructor(
-    private val taskApi: TaskApiService
-): RemoteDataSource {
-    override suspend fun getTask(taskId: String): Result<GetTaskResponse, DataError.Network> {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun postTask(task: TaskNetworkModel): EmptyResult<DataError> {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun updateTask(task: TaskNetworkModel): EmptyResult<DataError> {
-        TODO("Not yet implemented")
-    }
-
-    override suspend fun deleteTask(taskId: String): EmptyResult<DataError> {
-        TODO("Not yet implemented")
-    }
-}
